Making clean code

+++ BeautyServices.Service/DTOs/OrderDTo.cs
@@ -4,8 +4,8 @@ namespace BeautyServices.Service.DTOs;
 
 public class OrderDTo
 {
-    public int UserId { get; set; }
-    public int WorkerId { get; set; }
+    public long UserId { get; set; }
+    public long WorkerId { get; set; }
     public string Description { get; set; }
     public OrderTypes StatusType { get; set; }
 }
